rename "production plants" -> "generator systems"
