+++ packages/modules/io_actions/generator_systems/stepwise_control/api.py
@@ -5,7 +5,7 @@ from helpermodules.logger import ModifyLoglevelContext
 from modules.common.abstract_device import DeviceDescriptor
 from modules.common.abstract_io import AbstractIoAction
 from modules.common.utils.component_parser import get_component_name_by_id
-from modules.io_actions.production_plants.stepwise_control.config import StepwiseControlSetup
+from modules.io_actions.generator_systems.stepwise_control.config import StepwiseControlSetup
 
 control_command_log = logging.getLogger("steuve_control_command")
 
@@ -15,7 +15,7 @@ class StepwiseControl(AbstractIoAction):
         self.config = config
         control_command_log.info(f"Stufenweise Steuerung einer EZA: Eingang {self.config.configuration.s1} für S1, "
                                  f"Eingang {self.config.configuration.s2} für S2, und Eingang "
-                                 f"{self.config.configuration.w3} für W3 wird überwacht. Die Beschränkung musss in "
+                                 f"{self.config.configuration.w3} für W3 wird überwacht. Die Beschränkung muss in "
                                  "der EZA vorgenommen werden.")
         super().__init__()
 
@@ -24,7 +24,7 @@ class StepwiseControl(AbstractIoAction):
 
     def control_stepwise(self) -> Optional[str]:
         text = (f"Die Einspeiseleistung von {get_component_name_by_id(self.config.configuration.pv_id)} ist auf "
-                "{} % beschränkt. Die Beschränkung musss in der EZA vorgenommen werden.")
+                "{} % beschränkt. Die Beschränkung muss in der EZA vorgenommen werden.")
         msg = None
         digital_input = data.data.io_states[f"io_states{self.config.configuration.io_device}"].data.get.digital_input
         digital_input_prev = data.data.io_states[
@@ -55,12 +55,12 @@ class StepwiseControl(AbstractIoAction):
         else:
             # Keine Beschränkung soll nicht dauerhaft im WR angezeigt werden.
             msg = (f"Die Einspeiseleistung von {get_component_name_by_id(self.config.configuration.pv_id)} ist "
-                   "nicht beschränkt. Die Beschränkung musss in der EZA vorgenommen werden.")
+                   "nicht beschränkt. Die Beschränkung muss in der EZA vorgenommen werden.")
 
         if not (digital_input[self.config.configuration.s1] == digital_input_prev[self.config.configuration.s1] and
                 digital_input[self.config.configuration.s2] == digital_input_prev[self.config.configuration.s2] and
                 digital_input[self.config.configuration.w3] == digital_input_prev[self.config.configuration.w3]):
-            # Wenn sich was geändet hat, loggen
+            # Wenn sich was geändert hat, loggen
             with ModifyLoglevelContext(control_command_log, logging.DEBUG):
                 control_command_log.info(msg)
         return msg
