+++ packages/modules/io_actions/groups.py
@@ -3,10 +3,10 @@ from enum import Enum
 
 class ActionGroup(Enum):
     CONTROLLABLE_CONSUMERS = "controllable_consumers"
-    PRODUCTION_PLANTS = "production_plants"
+    GENERATOR_SYSTEMS = "generator_systems"
 
 
 READABLE_GROUP_NAME = {
     ActionGroup.CONTROLLABLE_CONSUMERS: "Steuerbare Verbrauchseinrichtungen (§14a)",
-    ActionGroup.PRODUCTION_PLANTS: "Erzeugungsanlagen (§9)",
+    ActionGroup.GENERATOR_SYSTEMS: "Erzeugungsanlagen (§9)",
 }
